+++ examples/src/index.js
@@ -3,13 +3,13 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Router, Route, browserHistory } from 'react-router';
 import App from './app';
-import Loader from './loading';
+import Loader from './loader';
 import './index.css';
 
 ReactDOM.render(
   <Router history={browserHistory}>
     <Route path="/" component={App} />
-    <Route path="loading" component={Loader} />
+    <Route path="/loading" component={Loader} />
   </Router>,
   document.getElementById('root')
 );
