+++ src/index.js
@@ -12,7 +12,7 @@ import Message from './components/message';
 import MessageInput from './components/message-input';
 import ProfileCard from './components/profile-card';
 import EmptyState from './components/empty-state';
-import Loading from './components/loading';
+import Loader from './components/loader';
 // Icons
 import IconAddImage from './icons/icon-add-image';
 import IconChannel from './icons/icon-channel';
@@ -54,7 +54,7 @@ export {
   MessageInput,
   ProfileCard,
   EmptyState,
-  Loading,
+  Loader,
   // Icons
   IconAddImage,
   IconChannel,
