+++ src/style/loader.js
